[#26207] YSQL: Ensure base row and unique index are flushed together in fast-path transaction
Summary:
Problem:

When running a COPY on a table with a unique index using a fast-path transaction, the COPY may fail
with inconsistent data in the database. In the example below, two rows are being copied to the same
table with a unique index on 'v', and both rows have a duplicate values on the column 'v'.

k              v
----------------
k1            v1
k2            v1
There are 4 writes: main table write for k1 and index table write for v1,
similarly 2 more for the next row. In the scenario, YSQL's operation
buffering logic will split the writes into two buffers due to duplicate
key on index v1. The first buffer will contain 3 entries:
(1) main table write of k1
(2) index table write for row1
(3) main table write of k1

The first buffer will be inserted into the database successfully; the second buffer (containing
1 entry: the index of row2) will fail, resulting in inconsistent data in the database.

Solution:

When a new write is detected to have duplicate key in the current buffer, it should be rejected if
2 conditions are met:

(1) The new write is from a fast-path txn.
(2) The new write is an INSERT.
Jira: DB-15553

Test Plan: New unit test:  ybd release --cxx-test pg_op_buffering-test --gtest_filter PgOpBufferingTest.BulkLoadForColocatedUseFastPathTxnTest

Reviewers: pjain, rthallam, dmitry, patnaik.balivada

Reviewed By: pjain, patnaik.balivada

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42391