[PLAT-16751][xCluster][dr] Retry-ability and abortability of xcluster/dr tasks
Summary:
This revision makes changes so the create/add/remove tables/dbs and replace dr replica tasks are retry-able.

Restart semantic prevents us from making it retryable. Restart is implemented to be able to run no matter the state of the xCluster/DR config and a restart task cannot identify if the current state is the result of a failing restart task (so it should continue from the latest subtask) or the result of a bad operation on the config (so restart must happen at the beginning of the task). We can put mechanisms to identify it but it complicates the task unnecessarily as a user should not need to restart the config often.

Test Plan: Aborted/retryed the create and add table tasks many times and made sure it works and can eventually completes. It was done for db-scoped and non-db-scoped DR configs with and without bootstrapping.

Reviewers: nsingh, amindrov, spothuraju, daniel, nbhatia, jmak

Reviewed By: amindrov

Differential Revision: https://phorge.dev.yugabyte.com/D42033