[#26332] docdb: Vector index compaction support in yb-admin
Summary:
The code base is updated to allow vector index compaction be triggered by the following command:
```
yb-admin compact_table <vector index name>
yb-admin compact_table_by_id <vector index id>
```
These commands trigger only vector index compaction and no compaction is triggered for the indexable
table. On the other hand, triggering compaction on the indexable table will not trigger compaction on
vector indexes -- this approach is selected to prevent undesirable effects of long vector index
compactions (because they are not yet optimized).

Additionally, `yb-admin flush_table <vector index>` is changed to trigger a flush only for vector
index and no flushes are triggered for the indexable table. Previously, the given command was
acting as if the flush was triggered for the indexable table.

**Upgrade/Rollback safety:**
It is safe to upgrade and rollback as the feature is not released. Also the presence of the values
are ignored by old releases which is acceptable behaviour (same is without the change).
Jira: DB-15678

Test Plan: Jenkins

Reviewers: sergei, slingam

Reviewed By: sergei

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D42404