[#26338] docdb: Vector index compaction support in yb-ts-cli
Summary:
The change implements new command `yb-ts-cli compact_vector_index` to allow particular tablet's
vector indexes compaction. The format:
```
 yb-ts-cli compact_vector_index <tablet_id> [<vector_index_id1> <vector_index_id12> ...]
```

**Upgrade/Rollback safety:**
It is safe to upgrade and rollback as the feature is not released. Also the presence of the values
are ignored by old releases which is acceptable behaviour (same is without the change).

Jira: DB-15687

Test Plan: Jenkins

Reviewers: sergei, slingam

Reviewed By: sergei

Subscribers: svc_phabricator, ybase

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D42413