[PLAT-16980] Rollback and retry for rollback of switchover doesn't work
Summary:
This diff fixes the annotation for the switchover rollback task as we do not support the rollback of a switchover rollback task.

It also adds the switchover rollback type so the task params can be created properly for the rollback to be retried.

Test Plan: Brought down one node on the dr replica, and ran switchover. It failed during network connectivity check and then could roll it back.

Reviewers: amindrov, spothuraju, jmak, yash.priyam, daniel

Reviewed By: jmak

Differential Revision: https://phorge.dev.yugabyte.com/D42427