[PLAT-16992] Install packages in isolated environment as fallback
Summary:
We were installing the pacakages in non isolated environments. In air gap environments, the pip version is older which does not support that option. This diff adds isolated environment installation as a fallback mechanism.
Also, fixes a minor issue with airgap installation.

Test Plan: Manually verified on the airgapped AMI.

Reviewers: skhilar, anijhawan

Reviewed By: anijhawan

Differential Revision: https://phorge.dev.yugabyte.com/D42442