[#26354] YSQL: fix regress linter to catch more cases
Summary:
In src/lint/regress_test.sh, the pattern for finding diff hunk markers
is too strict and fails to catch single-line hunks.  For example,

    222c219

is not caught, but

    420,421d410

is.  Fix the pattern to handle both cases.
Jira: DB-15708

Test Plan:
On Almalinux 8, before this change,

    src/lint/regress_test.sh src/postgres/src/test/regress/sql/yb.port.partition_aggregate.sql

gives 22.  After this change, it gives 53, mainly due to /* and */
lines.

Close: #26354
Jenkins: skip

Reviewers: sanketh, #db-approvers, slingam

Reviewed By: sanketh, #db-approvers, slingam

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42464