[#23785] YSQL: Incrementally refresh PG backend catcaches (part 6)
Summary:
In this diff, I added code changes to support interop between different releases.
For each invalidation message type, I added a version field. The idea is that
when a version matches, a message generated by a PG-backend of a another release
can be applied by this PG-backend. If a release introduces changes in
invalidation messages that cause interop with an old release no longer possible,
the version field will have a newer value to trigger a version mismatch. When
the old or new release detects a version mismatch, it will not do the
incremental catalog cache refresh optimization and will fall back to full
catalog cache refresh.

A new YB specific restriction is that new catcache IDs should be added at the
end of the `SysCacheIdentifier` enum list so that the existing catcache IDs can
remain the same. A new function `YbCheckCatalogCacheIds` is added which uses
static assert to detect any change in the `SysCacheIdentifier` enum list. If the
assertion fails, we need to evaluate whether `YbSharedInvalCatcacheMsgVersion`
needs to be incremented or not.

If an existing ID is removed, interop isn't possible so we need to
bump YbSharedInvalCatcacheMsgVersion.

If new ids are added, we need to add them at the end of the above
list. This is to allow we can keep `YbSharedInvalCatcacheMsgVersion`
unchanged so that old release PG backend can apply messages of any
existing ids. If a message catcache id's integer value is out of bound,
it will not be applied because the corresponding catalog cache does not
exist in the old PG backend. If there is a new id appended, we need to
evaluate whether we need to bump `YbSharedInvalCatcacheMsgVersion` or not:

(1) if the new id represents a new catalog table that did not exist in
in the old release, then interop is possible when the new id is not
involved, and there is no need to bump `YbSharedInvalCatcacheMsgVersion`.

(2) if the new id represents an existing catalog table in the old
release, then in a new PG backend the new catcache should be invalidated
but old PG backend cannot provide that message needed. In this case
interop isn't possible so we need to bump `YbSharedInvalCatcacheMsgVersion`.

Some unit tests are modified to allow them to pass no matter whether the gflag
`--TEST_yb_enable_invalidation_messages` is enabled or not.

Test Plan: ./yb_build.sh --cxx-test pg_catalog_version-test

Reviewers: kfranz, sanketh, mihnea

Reviewed By: kfranz

Subscribers: yql, svc_phabricator

Differential Revision: https://phorge.dev.yugabyte.com/D42468