[#25096] YSQL: Wait using StringWaiterLogSink in pg_local_limit_optimization-test
Summary:
### Issue

PgLocalLimitOptimizationTest is flaky with the following failure

```
Expected equality of these values:
  rows.size()
    Which is: 100002
  kNumInitialRows + 1
    Which is: 100001
```

### Cause

The test runs a select query concurrently with an insert statement, but the timing of the INSERT is critical. If the insert runs before the read, then the operations are not concurrent, leading to an incorrect scenario for the test. To correctly verify local limit behavior, the INSERT must occur after the read time to trigger a read restart error, something the local limit mechanism prevents.

The test currently checks for concurrency by verifying the row count. If the read sees the inserted row, it means the operations were not concurrent, causing the test to fail. The prior approach relies on sleeping before INSERT, but this is unreliable, as the INSERT may sometimes run too early, leading to flakiness.

### Resolution

To eliminate this uncertainty, this revision replaces the sleep with a wait mechanism based on a specific log line. The test now uses StringWaiterLogSink to wait for a log entry indicating when the SELECT query picks the read time. By setting `vmodule=read_query=1`, the waiter ensures that INSERT executes only after the `Read time` log appears, guaranteeing concurrency with SELECT.
Jira: DB-14242

Test Plan:
Jenkins

Ran the test `*PgLocalLimitOptimizationTest.*` 50 times on jenkins using `INSTANTIATE_TEST_SUITE_P(, PgLocalLimitOptimizationTest, ::testing::Range(1, 51))`

Reviewers: sergei, pjain, smishra

Reviewed By: sergei

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D41170