[PLAT-16844] remove lock which hinders parallel download from universe nodes
Summary:
We were acquiring a lock on universe UUID before downloading files from nodes of a universe.
This synchronises the threads and we lose all the benefits of parallelization.
There is no shared state being modified so the lock is not needed and hence removed.

Test Plan:
Manually verified faster task completion on local universe.
Verified time improvement from >13 mins to < 3 mins on the test cluster where we noticed this issue.

Reviewers: skurapati

Reviewed By: skurapati

Differential Revision: https://phorge.dev.yugabyte.com/D42479