[PLAT-16591] YBA Health checks for earlyoom
Summary:
Added a health check to verify that earlyoom is working as expected.
If earlyoom is available we are also checking mlock ability (and show warning if cannot)
Also there is a metric to count a number of kills by earlyoom during last hour.

Test Plan:
1) create universe with earlyoom enabled but without mlock ability
2) verify that we have warning about mlock
3) kill earlyoom - verify we have error
4) check number of kills metrics

create universe without earlyoom - see no alerts

Reviewers: anijhawan

Reviewed By: anijhawan

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D43339