[#3109] YSQL: Unify CommitPlainTransaction and CommitPlainTransactionContainingDDL from PgGate
Summary:
As part of D41998, the MVP support for the unification of DDL and DML transactions was added.

I missed one of the review comments which requested to unify the functions CommitPlainTransaction and CommitPlainTransactionContainingDDL from PgGate onwards. This revision tackles that.
Jira: DB-1839

Test Plan: `./yb_build.sh --java-test 'org.yb.pgsql.TestDdlTransactionBlocks'`

Reviewers: myang, pjain, skumar

Reviewed By: myang

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42485