[PLAT-16911] Introducing full move for the case of changing image bundle uuid in edit flow
Summary:
There is a case when our ami upgrade flow is not applicable (if ami is lost). So we can do full move to switch to a new ami.
Current diff implements that ability.
Due to the fact, that we don't write user intent to the db before update succeeds, we need to propagate imageBundleUUID through params (otherwise NodeManager uses the old one)

Test Plan:
1) create universe
2) run universe configure call with modified ami
3) run edit universe
4) verify new servers are using new ami

Reviewers: svarshney, spothuraju

Reviewed By: svarshney, spothuraju

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D42607