[PLAT-16924] Merge cloudInfo fields prior to making edit provider calls from the UI
Summary:
Not all readonly/unexposed cloudInfo fields are merged on the backend side. There is an
expectation that none of these fields are missing when editing otherwise it will be
perceived as an edit by the backend (unless explictly handled by computing the field
like in the case of AWS hosted zone name).

To make the contract between API and UI straightforward, the client will just pass along any field that it does
not deal with directly.

Test Plan: Successful edit provider operation for in-use AWS, GCP, AZU, K8s and on-prem provider.

Reviewers: svarshney

Reviewed By: svarshney

Differential Revision: https://phorge.dev.yugabyte.com/D42638