[#26419] YSQL: Introduce custom ParameterStatus packets for Connection Manager
Summary: This diff introduces a custom ParameterStatus (ConnMgrParameterStatus) packet to send from the server to with Connection Manager. The purpose of this ConnMgrParameterStatus packet is to signal to Connection Manager to not send back the packet to the client, while still functioning like a ParameterStatus packet to be absorbed by Connection Manager. This in turn will allow Connection Manager to maintain protocol level correctness with respect to how PG operates, rather than bombarding the client with ParameterStatus packets which get sent specifically for Connection Manager to store the GUC state of a client.

Test Plan: Jenkins: all tests, enable connection manager

Reviewers: skumar, mkumar, devansh.saxena

Reviewed By: skumar

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D42515