[PLAT-5231] Add new query info fields available in Postgres 15 to YBA UI
Summary:
YBA API exposes new fields for slow queries and live queries after Postgres 15 upgrade.
See:
6a7df549709995020beeb90c4e34fcd27446eaa8
2d36243d538019756d4ca4af27029d016eeb7f88
for related backend changes.

This diff exposes the new fields to the user on the universe `queries` tab.

The change uses the YugabyteDB software version as an indicator for whether
PG 15 is supported. In subsequent API updates, we will add the PG version directly
in the response payload so that clients are able to know immediately what PG version
a particular slow query/live query response is based on.

Test Plan:
Create a universe and upgrade to Postgres 15.
View queries tab and verify the expected fields are shown.
{F339390}
{F339441}
{F339449}

Reviewers: asharma, rmadhavan, lsangappa, kkannan

Reviewed By: rmadhavan

Differential Revision: https://phorge.dev.yugabyte.com/D42373